# openapi.yaml
openapi: 3.0.3
info:
  title: Bitcoin Blockchain Explorer API
  description: >
    Versioned JSON API under `/api/v1`. Hand-maintained; mirrors route registration in
    `internal/server/routes.go`. Errors use a flat envelope (code, message, correlation_id,
    timestamp RFC3339). List `pagination` objects use `total`, `page`, `page_size`, `has_more`.
    Deprecation: docs/API_VERSIONING.md. Session auth uses the `session_id`
    cookie; state-changing
    requests with that cookie require `X-CSRF-Token` (see `csrfMiddleware` in the server).
    Correlation IDs propagate via `X-Correlation-ID` / `X-Request-ID`.
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: "{baseUrl}"
    variables:
      baseUrl:
        default: http://localhost:8080
        description: Origin (no trailing slash)

# Default: no API key or cookie required; override with `security` on protected operations.
security: []
...
paths:
...
  /api/v1/blocks/export/csv:
    get:
      tags: [explorer]
      summary: Export blocks as CSV
      operationId: exportBlocksCSV
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: CSV stream
          content:
            text/csv:
              schema:
                type: string
        default:
          $ref: "#/components/responses/ApiError"
  /api/v1/transactions/export/csv:
    get:
      tags: [explorer]
      summary: Export transactions as CSV
      operationId: exportTransactionsCSV
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: CSV stream
          content:
            text/csv:
              schema:
                type: string
        default:
          $ref: "#/components/responses/ApiError"
...
  /api/v1/network-status:
    get:
      tags: [explorer]
      summary: Network status snapshot
      operationId: getNetworkStatus
      responses:
        "200":
          description: Network status JSON
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericJSON"
        default:
          $ref: "#/components/responses/ApiError"
...
  /api/v1/user/portfolios:
    get:
      tags: [user]
      summary: List portfolios
      operationId: listPortfolios
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: sort_by
          in: query
          schema:
            type: string
        - name: sort_dir
          in: query
          schema:
            type: string
            enum: [asc, desc]
      responses:
        "200":
          description: Portfolio list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericJSON"
        "401":
          $ref: "#/components/responses/ApiError"
        default:
          $ref: "#/components/responses/ApiError"
    post:
      tags: [user]
      summary: Create portfolio
      operationId: createPortfolio
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/CsrfHeader"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenericJSON"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericJSON"
        "401":
          $ref: "#/components/responses/ApiError"
        default:
          $ref: "#/components/responses/ApiError"
...
  /api/v1/user/watchlists/{id}:
    get:
      tags: [user]
      summary: Get watchlist by id
      operationId: getWatchlist
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Watchlist detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericJSON"
        "401":
          $ref: "#/components/responses/ApiError"
        default:
          $ref: "#/components/responses/ApiError"
    put:
      tags: [user]
      summary: Update watchlist
      operationId: updateWatchlist
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/CsrfHeader"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenericJSON"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericJSON"
        "401":
          $ref: "#/components/responses/ApiError"
        default:
          $ref: "#/components/responses/ApiError"
    delete:
      tags: [user]
      summary: Delete watchlist
      operationId: deleteWatchlist
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/CsrfHeader"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericJSON"
        "401":
          $ref: "#/components/responses/ApiError"
        default:
          $ref: "#/components/responses/ApiError"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session_id
      description: Session established via POST /api/v1/login
  parameters:
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Optional; export endpoints may de-duplicate identical replays (see docs/IDEMPOTENCY_KEYS.md)
      schema:
        type: string
    CsrfHeader:
      name: X-CSRF-Token
      in: header
      required: false
      description: >
        Required when `session_id` is sent and the request is state-changing (POST/PUT/PATCH/DELETE)
        or targets `/api/v1/admin/*`. Browser clients receive the token from login/profile responses.

      schema:
        type: string
  schemas:
    ApiErrorBody:
      type: object
      required: [code, message, correlation_id, timestamp]
      properties:
        code:
          type: string
        message:
          type: string
        correlation_id:
          type: string
          description: Echo of request correlation (may be empty only if middleware did not run)
        timestamp:
          type: string
          format: date-time
          description: Error emission time (UTC, RFC3339)
    GenericJSON:
      type: object
      additionalProperties: true
      description: Application-specific JSON (success envelopes vary by route)
...
  responses:
    ApiError:
      description: Standard API error envelope (four required fields)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorBody"